Add doc comments.

+++ Sudoku.Core/Data/Meta/Grid.cs
@@ -7,14 +7,56 @@ using Sudoku.Data.Extensions;
 
 namespace Sudoku.Data.Meta
 {
+	/// <summary>
+	/// Encapsulates a sudoku grid.
+	/// </summary>
 	public sealed class Grid : ICloneable<Grid>, IEnumerable<short>, IEquatable<Grid>, IFormattable
 	{
+		/// <summary>
+		/// Indicates an empty grid, where all values are zero.
+		/// </summary>
 		public static readonly Grid Empty = new Grid();
 
 
+		/// <summary>
+		/// Binary masks of all cells.
+		/// </summary>
+		/// <remarks>
+		/// <para>This array stores binary representation of decimals.</para>
+		/// <para>
+		/// There are 81 cells in a sudoku grid, so this data structure uses
+		/// an array of size 81. Each element is a <see cref="short"/> value,
+		/// where the lower 9 bits indicates if the digit 1 to 9 exists or not.
+		/// If the value is <c>true</c> (i.e. binary is for 1), this digit will
+		/// <b>not</b> be exist. The higher 3 bits indicates the cell status. The
+		/// cases are below:
+		/// <list type="table">
+		/// <item>
+		/// <term>0b001 (1)</term>
+		/// <description>The cell is <see cref="CellStatus.Empty"/>.</description>
+		/// </item>
+		/// <item>
+		/// <term>0b010 (2)</term>
+		/// <description>The cell is <see cref="CellStatus.Modifiable"/>.</description>
+		/// </item>
+		/// <item>
+		/// <term>0b100 (4)</term>
+		/// <description>The cell is <see cref="CellStatus.Given"/>.</description>
+		/// </item>
+		/// </list>
+		/// </para>
+		/// </remarks>
+		/// <seealso cref="CellStatus"/>
 		private readonly short[] _masks;
 
 
+		/// <summary>
+		/// Initializes an instance with the binary mask array.
+		/// </summary>
+		/// <param name="masks">The mask array.</param>
+		/// <exception cref="ArgumentException">
+		/// Throws when the length of the specified argument is not 81.
+		/// </exception>
 		public Grid(short[] masks)
 		{
 			if (masks.Length != 81)
@@ -29,6 +71,9 @@ namespace Sudoku.Data.Meta
 			ValueChanged += OnValueChanged;
 		}
 
+		/// <summary>
+		/// Provides default initialization.
+		/// </summary>
 		private Grid()
 		{
 			// 512 is equal to binary number '0b01_000_000_000', where the higher 2 bits
@@ -36,7 +81,7 @@ namespace Sudoku.Data.Meta
 			_masks = new short[81];
 			for (int i = 0; i < 81; i++)
 			{
-				_masks[i] = (int)CellStatus.Empty << 9;
+				_masks[i] = (short)CellStatus.Empty << 9;
 			}
 
 			// Initializes the event handler.
@@ -48,6 +93,10 @@ namespace Sudoku.Data.Meta
 		}
 
 
+		/// <summary>
+		/// Indicates the grid has already solved. If the value is <c>true</c>,
+		/// the grid is solved; otherwise, <c>false</c>.
+		/// </summary>
 		public bool HasSolved
 		{
 			get
@@ -89,6 +138,7 @@ namespace Sudoku.Data.Meta
 						}
 					}
 
+					// Each unique puzzle has at least 17 hints.
 					return count >= 17;
 				}
 				return SimplyValidate();
@@ -96,6 +146,22 @@ namespace Sudoku.Data.Meta
 		}
 
 
+		/// <summary>
+		/// Gets or sets a digit into a cell.
+		/// </summary>
+		/// <param name="offset">The cell offset you want to get or set.</param>
+		/// <value>
+		/// The digit you want to set. This value should be between 0 and 8.
+		/// In addtition, if your input is -1, the candidate mask in this cell
+		/// will be re-computed. If your input is none of them above, this indexer
+		/// will do nothing.
+		/// </value>
+		/// <returns>
+		/// An <see cref="int"/> value indicating the result.
+		/// If the current cell does not have a digit
+		/// (i.e. The cell is <see cref="CellStatus.Empty"/>),
+		/// The value will be -1.
+		/// </returns>
 		public int this[int offset]
 		{
 			get
@@ -128,7 +194,7 @@ namespace Sudoku.Data.Meta
 					short copy = result;
 
 					// Set cell status to 'CellStatus.Modifiable'.
-					result = (short)((int)CellStatus.Modifiable << 9 | 511 & ~(1 << value));
+					result = (short)((short)CellStatus.Modifiable << 9 | 511 & ~(1 << value));
 
 					// To trigger the event, which is used for eliminate
 					// all same candidates in peer cells.
@@ -182,6 +248,16 @@ namespace Sudoku.Data.Meta
 			}
 		}
 
+		/// <summary>
+		/// Gets or sets a candidate existence case with a <see cref="bool"/> value.
+		/// </summary>
+		/// <param name="offset">The cell offset between 0 and 80.</param>
+		/// <param name="digit">The digit between 0 and 8.</param>
+		/// <value>
+		/// The case you want to set. <c>true</c> means that this candidate
+		/// does not exist in this current sudoku grid; otherwise, <c>false</c>.
+		/// </value>
+		/// <returns>A <see cref="bool"/> value indicating that.</returns>
 		public bool this[int offset, int digit]
 		{
 			get => (_masks[offset] >> digit & 1) != 0;
@@ -204,9 +280,16 @@ namespace Sudoku.Data.Meta
 		}
 
 
+		/// <summary>
+		/// Indicates the event when the mask in a certain cell has changed.
+		/// </summary>
 		public event ValueChangedEventHandler ValueChanged;
 
 
+		/// <summary>
+		/// To fix a grid, which means all modifiable values will be changed
+		/// to given ones.
+		/// </summary>
 		public void FixGrid()
 		{
 			for (int i = 0; i < 81; i++)
@@ -218,6 +301,10 @@ namespace Sudoku.Data.Meta
 			}
 		}
 
+		/// <summary>
+		/// To unfix a grid, which means all given values will be changed
+		/// to modifiable ones.
+		/// </summary>
 		public void UnfixGrid()
 		{
 			for (int i = 0; i < 81; i++)
@@ -229,6 +316,11 @@ namespace Sudoku.Data.Meta
 			}
 		}
 
+		/// <summary>
+		/// Set the status in a cell.
+		/// </summary>
+		/// <param name="offset">The cell offset you want to change.</param>
+		/// <param name="cellStatus">The cell status you want to set.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public void SetCellStatus(int offset, CellStatus cellStatus)
 		{
@@ -239,6 +331,11 @@ namespace Sudoku.Data.Meta
 			ValueChanged.Invoke(this, new ValueChangedEventArgs(offset, copy, mask, -1));
 		}
 
+		/// <summary>
+		/// Set a mask in a cell.
+		/// </summary>
+		/// <param name="offset">The cell offset you want to change.</param>
+		/// <param name="value">The cell mask you want to set.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public void SetMask(int offset, short value)
 		{
@@ -249,11 +346,21 @@ namespace Sudoku.Data.Meta
 			ValueChanged.Invoke(this, new ValueChangedEventArgs(offset, copy, mask, -1));
 		}
 
+		/// <inheritdoc/>
 		public override bool Equals(object? obj) =>
 			obj is Grid comparer && Equals(comparer);
 
+		/// <summary>
+		/// Indicates whether the current object has the same value with the other one.
+		/// </summary>
+		/// <param name="other">The other value to compare.</param>
+		/// <returns>
+		/// The result of this comparsion. <c>true</c> if two instances hold a same
+		/// value; otherwise, <c>false</c>.
+		/// </returns>
 		public bool Equals(Grid other) => GetHashCode() == other.GetHashCode();
 
+		/// <inheritdoc/>
 		public override int GetHashCode()
 		{
 			int result = GetType().GetHashCode() ^ nameof(_masks).GetHashCode();
@@ -266,6 +373,13 @@ namespace Sudoku.Data.Meta
 			return result;
 		}
 
+		/// <summary>
+		/// Serializes this instance to an array, where all digit value will be stored.
+		/// </summary>
+		/// <returns>
+		/// This array. All elements are between 0 to 9, where 0 means the
+		/// cell is <see cref="CellStatus.Empty"/> now.
+		/// </returns>
 		public int[] ToArray()
 		{
 			int[] result = new int[81];
@@ -279,13 +393,26 @@ namespace Sudoku.Data.Meta
 			return result;
 		}
 
+		/// <summary>
+		/// Get a mask of the specified cell.
+		/// </summary>
+		/// <param name="offset">The cell offset you want to get.</param>
+		/// <returns>The mask.</returns>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public short GetMask(int offset) => _masks[offset];
 
+		/// <inheritdoc/>
 		public override string ToString() => ToString(null, null);
 
+		/// <summary>
+		/// Returns a string that represents the current object, with the
+		/// specified format.
+		/// </summary>
+		/// <param name="format">The format.</param>
+		/// <returns>The string result.</returns>
 		public string ToString(string format) => ToString(format, null);
 
+		/// <inheritdoc/>
 		public string ToString(string? format, IFormatProvider? formatProvider)
 		{
 			if (formatProvider?.GetFormat(GetType()) is ICustomFormatter customFormatter)
@@ -593,17 +720,30 @@ namespace Sudoku.Data.Meta
 			}
 		}
 
+		/// <summary>
+		/// Get a cell status of the specified cell.
+		/// </summary>
+		/// <param name="offset">The cell offset you want to get.</param>
+		/// <returns>The cell status.</returns>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public CellStatus GetCellStatus(int offset) =>
 			(CellStatus)(_masks[offset] >> 9 & (int)CellStatus.All);
 
+		/// <inheritdoc/>
 		public Grid Clone() => new Grid((short[])_masks.Clone());
 
+		/// <inheritdoc/>
 		public IEnumerator<short> GetEnumerator() =>
 			((IEnumerable<short>)_masks).GetEnumerator();
 
+		/// <inheritdoc/>
 		IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
+		/// <summary>
+		/// The method, which will be invoked when the mask has changed.
+		/// </summary>
+		/// <param name="sender">The instance triggering the event.</param>
+		/// <param name="e">The data.</param>
 		private void OnValueChanged(object sender, ValueChangedEventArgs e)
 		{
 			var (offset, _, _, setValue) = e;
@@ -628,8 +768,23 @@ namespace Sudoku.Data.Meta
 		}
 
 
+		/// <summary>
+		/// Parses a string value and converts to this type.
+		/// </summary>
+		/// <param name="str">The string.</param>
+		/// <returns>The result instance had converted.</returns>
 		public static Grid Parse(string str) => new GridParser(str).Parse();
 
+		/// <summary>
+		/// Try to parse a string and converts to this type, and returns a
+		/// <see cref="bool"/> value indicating the result of the conversion.
+		/// </summary>
+		/// <param name="str">The string.</param>
+		/// <param name="result">
+		/// (out parameter) The result parsed. If the conversion is failed,
+		/// this argument will be <c>null</c>.
+		/// </param>
+		/// <returns>A <see cref="bool"/> value indicating that.</returns>
 		public static bool TryParse(string str, [NotNullWhen(true)] out Grid? result)
 		{
 			try
@@ -644,6 +799,11 @@ namespace Sudoku.Data.Meta
 			}
 		}
 
+		/// <summary>
+		/// Creates an instance using grid values.
+		/// </summary>
+		/// <param name="gridValues">The array of grid values.</param>
+		/// <returns>The result instance.</returns>
 		public static Grid CreateInstance(int[] gridValues)
 		{
 			var result = Empty.Clone();
@@ -664,6 +824,11 @@ namespace Sudoku.Data.Meta
 			return result;
 		}
 
+		/// <summary>
+		/// Creates an instance using grid values.
+		/// </summary>
+		/// <param name="gridValues">The array of grid values.</param>
+		/// <returns>The result instance.</returns>
 		public static Grid CreateInstance(int[,] gridValues)
 		{
 			var result = Empty.Clone();
@@ -685,8 +850,20 @@ namespace Sudoku.Data.Meta
 		}
 
 
+		/// <summary>
+		/// Indicates whether two instances have a same value.
+		/// </summary>
+		/// <param name="left">The left instance.</param>
+		/// <param name="right">The right instance.</param>
+		/// <returns>A <see cref="bool"/> result indicating that.</returns>
 		public static bool operator ==(Grid left, Grid right) => left.Equals(right);
 
+		/// <summary>
+		/// Indicates whether two instances have two different values.
+		/// </summary>
+		/// <param name="left">The left instance.</param>
+		/// <param name="right">The right instance.</param>
+		/// <returns>A <see cref="bool"/> result indicating that.</returns>
 		public static bool operator !=(Grid left, Grid right) => !(left == right);
 	}
 }
